# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: CoreBanking API
  version: 1.0.0-phase1
  description: |
    Staff and integration API for CoreBanking. Every request is tenant-scoped: the tenant comes from the
    verified access token (`tenant` claim), never from the URL or body.

    Conventions
    - Money is a decimal **string** (e.g. "100000.00"). Never a JSON number.
    - Dates are ISO `YYYY-MM-DD` business dates; timestamps are RFC 3339 UTC.
    - Changes to master data and money go through maker-checker: the call returns **202** with an
      approval request; the change is applied when a different user approves it.
    - Errors use RFC 9457 `application/problem+json`.
    - `Idempotency-Key` header is honoured on all POSTs that create things.
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local dev
security:
  - bearer: []
...
paths:
...
  /api/v1/customers:
    get:
      tags: [Customers]
      operationId: searchCustomers
      summary: Search customers
      parameters:
        - { name: q, in: query, description: 'Customer number, exact PAN or exact mobile', schema: { type: string } }
        - { name: externalRef, in: query, description: 'The id of the customer in the LOS (create-or-get): returns that customer or an empty list', schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, default: 0 } }
        - { name: size, in: query, schema: { type: integer, default: 20, maximum: 100 } }
      responses:
        '200':
          description: Masked customer summaries
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CustomerSummary' }
        default: { $ref: '#/components/responses/Problem' }
    post:
      tags: [Customers]
      operationId: createCustomer
      summary: Create a customer (maker-checker). Returns the existing customer when PAN already exists (idempotent create).
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CustomerInput' }
      responses:
        '200':
          description: Customer with this PAN already exists
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CustomerSummary' }
        '202': { $ref: '#/components/responses/Accepted' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/customers/dedupe-check:
    post:
      tags: [Customers]
      operationId: dedupeCheck
      summary: Dedupe check
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CustomerInput' }
      responses:
        '200':
          description: Possible duplicates with the matching rule
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/DedupeMatch' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/customers/{id}:
    get:
      tags: [Customers]
      operationId: getCustomer
      summary: Get customer
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses:
        '200':
          description: Masked customer
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CustomerSummary' }
        default: { $ref: '#/components/responses/Problem' }
...
  /api/v1/loans/{id}/collection-orders:
    post:
      tags: [Integrations]
      operationId: createCollectionOrder
      summary: Create a payment link or order for a loan
      description: 'Permission: collection:create.'
      parameters:
        - { $ref: '#/components/parameters/Id' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CollectionOrderInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IntegrationRecord' }
        default: { $ref: '#/components/responses/Problem' }
    get:
      tags: [Integrations]
      operationId: listCollectionOrders
      summary: Payment orders of a loan
      description: 'Permission: collection:view.'
      parameters:
        - { $ref: '#/components/parameters/Id' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/IntegrationRecord' } }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/collection-orders/{id}:
    get:
      tags: [Integrations]
      operationId: getCollectionOrder
      summary: A payment order
      description: 'Permission: collection:view.'
      parameters:
        - { $ref: '#/components/parameters/Id' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IntegrationRecord' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/gateway-payments/reconciliation:
    get:
      tags: [Integrations]
      operationId: getCollectionReconciliation
      summary: Gateway settlements against postings, unmatched in both directions
      description: 'Permission: collection:view.'
      parameters:
        - { name: category, in: query, description: 'MATCHED, PAYMENT_NOT_POSTED, POSTED_NOT_SETTLED, SETTLED_NOT_RECEIVED, AMOUNT_MISMATCH or REFUND_DUE; without it everything that is not MATCHED', schema: { type: string } }
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/IntegrationRecord' } }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/gateway-payments/{id}/resolve:
    post:
      tags: [Integrations]
      operationId: resolveGatewayPayment
      summary: Unmatched receipts queue - assign a receipt to a loan or mark it for refund
      description: 'Permission: collection:admin.'
      parameters:
        - { $ref: '#/components/parameters/Id' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PaymentResolution' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IntegrationRecord' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/gateway-settlements/upload:
    post:
      tags: [Integrations]
      operationId: uploadGatewaySettlements
      summary: Load a settlement report (CSV with provider_payment_id, amount, settled_on and optionally fee, utr)
      description: 'Permission: collection:admin.'
      parameters:
        - { name: fileRef, in: query, required: true, schema: { type: string, maxLength: 60 } }
      requestBody:
        required: true
        content:
          text/csv:
            schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IntegrationRecord' }
        default: { $ref: '#/components/responses/Problem' }
...
  /api/v1/api-clients:
    get:
      tags: [Integrations]
      operationId: listApiClients
      summary: OAuth2 API clients of the tenant
      description: 'Permission: apiclient:view.'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/IntegrationRecord' } }
        default: { $ref: '#/components/responses/Problem' }
    post:
      tags: [Integrations]
      operationId: proposeApiClient
      summary: Propose an API client (maker-checker; two checkers when a money-moving scope is included)
      description: 'Permission: apiclient:admin.'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ApiClientInput' }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/api-clients/scopes:
    get:
      tags: [Integrations]
      operationId: listApiClientScopes
      summary: The scopes an API client may hold and which need two checkers
      description: 'Permission: apiclient:view.'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IntegrationRecord' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/api-clients/{clientId}:
    get:
      tags: [Integrations]
      operationId: getApiClient
      summary: An API client
      description: 'Permission: apiclient:view.'
      parameters:
        - { name: clientId, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IntegrationRecord' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/api-clients/{clientId}/scopes:
    put:
      tags: [Integrations]
      operationId: proposeApiClientScopes
      summary: Propose a scope change (maker-checker)
      description: 'Permission: apiclient:admin.'
      parameters:
        - { name: clientId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ApiClientScopes' }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/api-clients/{clientId}/actions:
    post:
      tags: [Integrations]
      operationId: proposeApiClientAction
      summary: Propose ROTATE_SECRET, DISABLE or ENABLE (maker-checker)
      description: 'Permission: apiclient:admin.'
      parameters:
        - { name: clientId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ActionInput' }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/api-clients/{clientId}/secret:
    post:
      tags: [Integrations]
      operationId: collectApiClientSecret
      summary: Collect the client secret - once, by the user who proposed the client or the rotation
      description: 'Permission: apiclient:admin. The secret is generated at the identity provider at this moment and is not stored by CoreBanking.'
      parameters:
        - { name: clientId, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IntegrationRecord' }
        default: { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      schema: { type: string, maxLength: 100 }
...
  responses:
    Accepted:
      description: Change recorded as an approval request; applied when a checker approves it
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Approval' }
    Problem:
      description: Error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Money:
      type: string
      pattern: '^-?[0-9]+(\.[0-9]{1,4})?$'
      examples: ['100000.00']
    Problem:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        errors:
          type: array
          description: Field-level problems (validation), when there are several
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
...
    ApprovalStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED, WITHDRAWN]
    Approval:
      type: object
      required: [id, entityType, action, status, maker, madeAt]
      properties:
        id: { type: string, format: uuid }
        entityType: { type: string, examples: [CUSTOMER, BRANCH, BRANCH_SET, STAFF, GL_HEAD, VOUCHER, TAX_RATE, HOLIDAY, TERRITORY, ENUMERATION, SYSTEM_PROPERTY, EOD_SCHEDULE, LOAN_PRODUCT, LOAN_DISBURSEMENT, LOAN_WAIVER, LOAN_REVERSAL, LOAN_AMENDMENT, LOAN_RESTRUCTURE] }
        entityId: { type: [string, 'null'] }
        action: { type: string, examples: [CREATE, UPDATE, REVERSE] }
        status: { $ref: '#/components/schemas/ApprovalStatus' }
        maker: { type: string }
        madeAt: { type: string, format: date-time }
        checker: { type: [string, 'null'] }
        checkedAt: { type: [string, 'null'], format: date-time }
        note: { type: [string, 'null'] }
        amount: { oneOf: [ { $ref: '#/components/schemas/Money' }, { type: 'null' } ] }
        ageHours: { type: number, description: SLA ageing }
        checkersRequired: { type: integer, description: 'Distinct checkers needed (0-3), from the approval rules' }
        approvalsSoFar: { type: integer, description: Approvals already recorded while PENDING }
        appliedRef: { type: [string, 'null'], description: 'What the approval produced (customer id, voucher number …)' }
        current: { type: [object, 'null'], description: State before the change (null for CREATE) }
        proposed: { type: object, description: Proposed state }
...
    CustomerInput:
      type: object
      required: [customerType, firstName, homeBranch, dateOfBirth, mobile]
      properties:
        custom: { $ref: '#/components/schemas/CustomValues' }
        customerType: { type: string, enum: [INDIVIDUAL, NON_INDIVIDUAL] }
        firstName: { type: string }
        middleName: { type: string }
        lastName: { type: string }
        dateOfBirth: { type: string, format: date }
        gender: { type: string, enum: [FEMALE, MALE, OTHER] }
        pan: { type: string, pattern: '^[A-Z]{5}[0-9]{4}[A-Z]$' }
        mobile: { type: string, pattern: '^[6-9][0-9]{9}$' }
        email: { type: string, format: email }
        homeBranch: { type: string }
        address:
          type: object
          properties:
            line1: { type: string }
            line2: { type: string }
            city: { type: string }
            stateCode: { type: string }
            pincode: { type: string, pattern: '^[1-9][0-9]{5}$' }
        overrideDedupe: { type: boolean, description: Proceed despite possible duplicates (requires approval with reason) }
        overrideReason: { type: string }
        externalRef: { type: string, pattern: '^[A-Za-z0-9._:-]{1,64}$', description: 'The id of the customer in the LOS; repeating it returns the existing customer' }
    CustomerSummary:
      type: object
      required: [id, customerNo, displayName, status]
      properties:
        custom: { $ref: '#/components/schemas/CustomValues' }
        id: { type: string, format: uuid }
        customerNo: { type: string }
        displayName: { type: string }
        customerType: { type: string }
        dateOfBirth: { type: string, format: date }
        panMasked: { type: [string, 'null'], examples: ['XXXXX1234X'] }
        mobileMasked: { type: [string, 'null'], examples: ['XXXXXX0001'] }
        homeBranch: { type: string }
        kycStatus: { type: string }
        status: { type: string }
    DedupeMatch:
      type: object
      required: [customerNo, displayName, rule]
      properties:
        customerNo: { type: string }
        displayName: { type: string }
        rule: { type: string, enum: [PAN, MOBILE, NAME_DOB] }
        strength: { type: string, enum: [EXACT, STRONG, POSSIBLE] }
...
    IntegrationRecord:
      type: object
      description: A record of the integration module. Money is a decimal string; account numbers appear only masked.
      additionalProperties: true
...
    CollectionOrderInput:
      type: object
      required: [amount]
      properties:
        amount: { $ref: '#/components/schemas/Money' }
        methods: { type: array, items: { type: string, enum: [UPI, CARD, NETBANKING] } }
        returnUrl: { type: string, format: uri }
    PaymentResolution:
      type: object
      required: [note]
      properties:
        loanId: { type: string, format: uuid }
        refund: { type: boolean }
        note: { type: string }
...
    ActionInput:
      type: object
      required: [action]
      properties:
        action: { type: string }
    ApiClientInput:
      type: object
      required: [clientId, name, scopes, homeBranch]
      properties:
        clientId: { type: string, description: 'Lower-case letters, digits and hyphens; stored with the prefix ext-' }
        name: { type: string, maxLength: 80 }
        scopes: { type: array, items: { type: string } }
        homeBranch: { type: string }
        allBranches: { type: boolean, default: false }
    ApiClientScopes:
      type: object
      required: [scopes]
      properties:
        scopes: { type: array, items: { type: string } }
...
    CustomValues:
      type: object
      description: >-
        The tenant's custom fields by key (see `GET /api/v1/custom-fields`). An unknown key, a value of the wrong type
        and a missing required field are refused with 422. Personal-data fields are returned masked.
      additionalProperties: true
...